+++ examples/1d_stencil/1d_stencil_4.cpp
@@ -203,10 +203,9 @@ struct stepper
 ///////////////////////////////////////////////////////////////////////////////
 int hpx_main(hpx::program_options::variables_map& vm)
 {
-    std::uint64_t np = vm["np"].as<std::uint64_t>();    // Number of partitions.
-    std::uint64_t nx =
-        vm["nx"].as<std::uint64_t>();    // Number of grid points.
-    std::uint64_t nt = vm["nt"].as<std::uint64_t>();    // Number of steps.
+    std::size_t np = vm["np"].as<std::size_t>();    // Number of partitions.
+    std::size_t nx = vm["nx"].as<std::size_t>();    // Number of grid points.
+    std::size_t nt = vm["nt"].as<std::size_t>();    // Number of steps.
     std::uint64_t nd =
         vm["nd"].as<std::uint64_t>();    // Max depth of dep tree.
 
@@ -250,13 +249,13 @@ int main(int argc, char* argv[])
     // clang-format off
     desc_commandline.add_options()
         ("results", "print generated results (default: false)")
-        ("nx", value<std::uint64_t>()->default_value(10),
+        ("nx", value<std::size_t>()->default_value(10),
          "Local x dimension (of each partition)")
-        ("nt", value<std::uint64_t>()->default_value(45),
+        ("nt", value<std::size_t>()->default_value(45),
          "Number of time steps")
         ("nd", value<std::uint64_t>()->default_value(10),
          "Number of time steps to allow the dependency tree to grow to")
-        ("np", value<std::uint64_t>()->default_value(10),
+        ("np", value<std::size_t>()->default_value(10),
          "Number of partitions")
         ("k", value<double>(&k)->default_value(0.5),
          "Heat transfer coefficient (default: 0.5)")
