+++ examples/1d_stencil/1d_stencil_8.cpp
@@ -641,7 +641,7 @@ stepper_server::space stepper_server::do_work(
 
 ///////////////////////////////////////////////////////////////////////////////
 void do_all_work(
-    std::uint64_t nt, std::uint64_t nx, std::uint64_t np, std::uint64_t nd)
+    std::size_t nt, std::size_t nx, std::size_t np, std::uint64_t nd)
 {
     std::vector<hpx::id_type> localities = hpx::find_all_localities();
     std::size_t nl = localities.size();    // Number of localities
@@ -720,10 +720,9 @@ void do_all_work(
 ///////////////////////////////////////////////////////////////////////////////
 int hpx_main(hpx::program_options::variables_map& vm)
 {
-    std::uint64_t nt = vm["nt"].as<std::uint64_t>();    // Number of steps.
-    std::uint64_t nx =
-        vm["nx"].as<std::uint64_t>();    // Number of grid points.
-    std::uint64_t np = vm["np"].as<std::uint64_t>();    // Number of partitions.
+    std::size_t nt = vm["nt"].as<std::size_t>();    // Number of steps.
+    std::size_t nx = vm["nx"].as<std::size_t>();    // Number of grid points.
+    std::size_t np = vm["np"].as<std::size_t>();    // Number of partitions.
     std::uint64_t nd =
         vm["nd"].as<std::uint64_t>();    // Max depth of dep tree.
 
@@ -744,12 +743,12 @@ int main(int argc, char* argv[])
     options_description desc_commandline;
     desc_commandline.add_options()(
         "results", "print generated results (default: false)")("nx",
-        value<std::uint64_t>()->default_value(10),
+        value<std::size_t>()->default_value(10),
         "Local x dimension (of each partition)")("nt",
-        value<std::uint64_t>()->default_value(45),
+        value<std::size_t>()->default_value(45),
         "Number of time steps")("nd", value<std::uint64_t>()->default_value(10),
         "Number of time steps to allow the dependency tree to grow to")("np",
-        value<std::uint64_t>()->default_value(10),
+        value<std::size_t>()->default_value(10),
         "Number of partitions")("k", value<double>(&k)->default_value(0.5),
         "Heat transfer coefficient (default: 0.5)")("dt",
         value<double>(&dt)->default_value(1.0),
